split up data_query.go
